@@ -1,31 +1,45 @@
 apply plugin: 'java'
 apply plugin: 'application'
 
-sourceCompatibility = '1.8'
+sourceCompatibility = JavaVersion.VERSION_11
+targetCompatibility = JavaVersion.VERSION_11
 
 mainClassName = 'io.github.madrigal.Main'
 
 repositories {
-    jcenter()
+    mavenCentral()
 }
 
 //create a single Jar with all dependencies
 // it will be located in build/libs/XXX-all.jar
-task fatJar(type: Jar) {
+
+tasks.register('uberJar', Jar) {
+    archiveClassifier = 'uber'
+    duplicatesStrategy(DuplicatesStrategy.EXCLUDE)
+
+    from sourceSets.main.output
+
+    dependsOn configurations.runtimeClasspath
+    from {
+        configurations.runtimeClasspath.findAll { it.name.endsWith('jar') }.collect { zipTree(it) }
+    }
     manifest {
         attributes 'Implementation-Title': 'Pulse Jar',
                 'Implementation-Version': '1.0',
                 'Main-Class': mainClassName
     }
-    baseName = project.name + '-all'
-    from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
-    with jar
 }
+
+
 dependencies {
-    testCompile 'junit:junit:4.12'
-    compile 'com.google.code.gson:gson:2.6.2'
-    compile 'com.amazonaws:aws-lambda-java-core:1.1.0'
-    compile 'com.amazonaws:aws-lambda-java-events:1.1.0'
-    compile 'javax.mail:mail:1.4.7'
-    compile 'com.amazonaws:aws-java-sdk-ses:1.11.23'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
+    implementation 'com.google.code.gson:gson:2.10.1'
+    implementation 'com.amazonaws:aws-lambda-java-core:1.2.2'
+    implementation 'com.amazonaws:aws-lambda-java-events:3.11.0'
+    implementation 'javax.annotation:javax.annotation-api:1.3.2'
+    implementation 'javax.mail:mail:1.4.7'
+    implementation 'com.amazonaws:aws-java-sdk-ses:1.11.23'
+    // Fix 'java.lang.NoClassDefFoundError: javax/xml/bind/DatatypeConverter'
+    implementation 'javax.xml.bind:jaxb-api:2.3.1'
+    implementation 'org.glassfish.jaxb:jaxb-runtime:4.0.0'
 }
